requirements.txt
python-dotenv==0.20.0
asyncio~=3.4.3
isdayoff~=1.0
fastapi~=0.78.0
pydantic~=1.9.1
uvicorn~=0.17.6